Move compiled code to a test file, add List.foldl
