+++ src/kernel/core/elm.h
@@ -8,6 +8,7 @@ Closure g_elm_core_Maybe_Just;
 
 Closure g_elm_core_Array_initialize;
 Closure g_elm_core_List_reverse;
+Closure g_elm_core_List_foldl;
 
 void* eval_elm_core_Result_Ok(void* args[]);
 void* eval_elm_core_Result_Err(void* args[]);
